+++ Shared/Review.swift
@@ -120,12 +120,12 @@ final class Review: DataItem {
         }
     }
 
-    static func review(with id: Int64, in moc: NSManagedObjectContext) -> Review? {
+    static func review(with id: Int, in moc: NSManagedObjectContext) -> Review? {
         let f = NSFetchRequest<Review>(entityName: "Review")
         f.returnsObjectsAsFaults = false
         f.includesSubentities = false
         f.fetchLimit = 1
-        f.predicate = NSPredicate(format: "serverId == %lld", id)
+        f.predicate = NSPredicate(format: "serverId == %d", id)
         return try! moc.fetch(f).first
     }
